+++ runners/core-construction-java/src/main/java/org/apache/beam/runners/core/construction/graph/SplittableParDoExpander.java
@@ -17,6 +17,7 @@
  */
 package org.apache.beam.runners.core.construction.graph;
 
+import com.google.auto.value.AutoValue;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.Map;
@@ -66,7 +67,7 @@ public class SplittableParDoExpander {
    * information is available to the runner if it chooses to inspect it.
    */
   public static TransformReplacement createSizedReplacement() {
-    return SizedReplacement.INSTANCE;
+    return SizedReplacement.builder().setDrain(false).build();
   }
 
   /**
@@ -91,13 +92,25 @@ public class SplittableParDoExpander {
    * .
    */
   public static TransformReplacement createTruncateReplacement() {
-    return TruncateReplacement.INSTANCE;
+    return SizedReplacement.builder().setDrain(true).build();
   }
 
   /** See {@link #createSizedReplacement()} for details. */
-  private static class SizedReplacement implements TransformReplacement {
+  @AutoValue
+  abstract static class SizedReplacement implements TransformReplacement {
 
-    private static final SizedReplacement INSTANCE = new SizedReplacement();
+    static Builder builder() {
+      return new AutoValue_SplittableParDoExpander_SizedReplacement.Builder();
+    }
+
+    abstract boolean isDrain();
+
+    @AutoValue.Builder
+    abstract static class Builder {
+      abstract Builder setDrain(boolean isDrain);
+
+      abstract SizedReplacement build();
+    }
 
     @Override
     public MessageWithComponents getReplacement(
@@ -209,14 +222,71 @@ public class SplittableParDoExpander {
           splitAndSize.setEnvironmentId(splittableParDo.getEnvironmentId());
           rval.getComponentsBuilder().putTransforms(splitAndSizeId, splitAndSize.build());
         }
+        PTransform.Builder newCompositeRoot =
+            splittableParDo
+                .toBuilder()
+                // Clear the original splittable ParDo spec and add all the new transforms as
+                // children.
+                .clearSpec()
+                .addAllSubtransforms(Arrays.asList(pairWithRestrictionId, splitAndSizeId));
 
         String processSizedElementsAndRestrictionsId =
             generateUniqueId(
                 transformId + "/ProcessSizedElementsAndRestrictions",
                 existingComponents::containsTransforms);
+        String processSizedElementsInputPCollectionId = splitAndSizeOutId;
+        if (isDrain()) {
+          String truncateAndSizeCoderId =
+              generateUniqueId(
+                  mainInputPCollection.getCoderId() + "/TruncateAndSize",
+                  existingComponents::containsCoders);
+          rval.getComponentsBuilder()
+              .putCoders(
+                  truncateAndSizeCoderId,
+                  ModelCoders.kvCoder(
+                      splitAndSizeOutCoderId, getOrAddDoubleCoder(existingComponents, rval)));
+          String truncateAndSizeOutId =
+              generateUniqueId(
+                  mainInputPCollectionId + "/TruncateAndSize",
+                  existingComponents::containsPcollections);
+
+          rval.getComponentsBuilder()
+              .putPcollections(
+                  truncateAndSizeOutId,
+                  PCollection.newBuilder()
+                      .setCoderId(truncateAndSizeCoderId)
+                      .setIsBounded(mainInputPCollection.getIsBounded())
+                      .setWindowingStrategyId(mainInputPCollection.getWindowingStrategyId())
+                      .setUniqueName(
+                          generateUniquePCollectonName(
+                              mainInputPCollection.getUniqueName() + "/TruncateAndSize",
+                              existingComponents))
+                      .build());
+          String truncateAndSizeId =
+              generateUniqueId(
+                  transformId + "/TruncateAndSize", existingComponents::containsTransforms);
+          {
+            PTransform.Builder truncateAndSize = PTransform.newBuilder();
+            truncateAndSize.putInputs(mainInputName, splitAndSizeOutId);
+            truncateAndSize.putAllInputs(sideInputs);
+            truncateAndSize.putOutputs("out", truncateAndSizeOutId);
+            truncateAndSize.setUniqueName(
+                generateUniquePCollectonName(
+                    splittableParDo.getUniqueName() + "/TruncateAndSize", existingComponents));
+            truncateAndSize.setSpec(
+                FunctionSpec.newBuilder()
+                    .setUrn(PTransformTranslation.SPLITTABLE_TRUNCATE_SIZED_RESTRICTION_URN)
+                    .setPayload(splittableParDo.getSpec().getPayload()));
+            truncateAndSize.setEnvironmentId(splittableParDo.getEnvironmentId());
+            rval.getComponentsBuilder().putTransforms(truncateAndSizeId, truncateAndSize.build());
+          }
+          newCompositeRoot.addSubtransforms(truncateAndSizeId);
+          processSizedElementsInputPCollectionId = truncateAndSizeOutId;
+        }
         {
           PTransform.Builder processSizedElementsAndRestrictions = PTransform.newBuilder();
-          processSizedElementsAndRestrictions.putInputs(mainInputName, splitAndSizeOutId);
+          processSizedElementsAndRestrictions.putInputs(
+              mainInputName, processSizedElementsInputPCollectionId);
           processSizedElementsAndRestrictions.putAllInputs(sideInputs);
           processSizedElementsAndRestrictions.putAllOutputs(splittableParDo.getOutputsMap());
           processSizedElementsAndRestrictions.setUniqueName(
@@ -234,20 +304,8 @@ public class SplittableParDoExpander {
                   processSizedElementsAndRestrictionsId,
                   processSizedElementsAndRestrictions.build());
         }
-
-        PTransform.Builder newCompositeRoot =
-            splittableParDo
-                .toBuilder()
-                // Clear the original splittable ParDo spec and add all the new transforms as
-                // children.
-                .clearSpec()
-                .addAllSubtransforms(
-                    Arrays.asList(
-                        pairWithRestrictionId,
-                        splitAndSizeId,
-                        processSizedElementsAndRestrictionsId));
+        newCompositeRoot.addSubtransforms(processSizedElementsAndRestrictionsId);
         rval.setPtransform(newCompositeRoot);
-
         return rval.build();
       } catch (IOException e) {
         throw new RuntimeException("Unable to perform expansion for transform " + transformId, e);
@@ -297,212 +355,5 @@ public class SplittableParDoExpander {
       i += 1;
     }
     return prefix + i;
-  }
-
-  /** See {@link #createTruncateReplacement} for details. */
-  private static class TruncateReplacement implements TransformReplacement {
-    private static final TruncateReplacement INSTANCE = new TruncateReplacement();
-
-    @Override
-    public MessageWithComponents getReplacement(
-        String transformId, ComponentsOrBuilder existingComponents) {
-      try {
-        MessageWithComponents.Builder rval = MessageWithComponents.newBuilder();
-
-        PTransform splittableParDo = existingComponents.getTransformsOrThrow(transformId);
-        ParDoPayload payload = ParDoPayload.parseFrom(splittableParDo.getSpec().getPayload());
-        // Only perform the expansion if this is a splittable DoFn.
-        if (payload.getRestrictionCoderId() == null || payload.getRestrictionCoderId().isEmpty()) {
-          return null;
-        }
-
-        String mainInputName = ParDoTranslation.getMainInputName(splittableParDo);
-        String mainInputPCollectionId = splittableParDo.getInputsOrThrow(mainInputName);
-        PCollection mainInputPCollection =
-            existingComponents.getPcollectionsOrThrow(mainInputPCollectionId);
-        Map<String, String> sideInputs =
-            Maps.filterKeys(
-                splittableParDo.getInputsMap(), input -> payload.containsSideInputs(input));
-
-        String pairWithRestrictionOutCoderId =
-            generateUniqueId(
-                mainInputPCollection.getCoderId() + "/PairWithRestriction",
-                existingComponents::containsCoders);
-        rval.getComponentsBuilder()
-            .putCoders(
-                pairWithRestrictionOutCoderId,
-                ModelCoders.kvCoder(
-                    mainInputPCollection.getCoderId(), payload.getRestrictionCoderId()));
-
-        String pairWithRestrictionOutId =
-            generateUniqueId(
-                mainInputPCollectionId + "/PairWithRestriction",
-                existingComponents::containsPcollections);
-        rval.getComponentsBuilder()
-            .putPcollections(
-                pairWithRestrictionOutId,
-                PCollection.newBuilder()
-                    .setCoderId(pairWithRestrictionOutCoderId)
-                    .setIsBounded(mainInputPCollection.getIsBounded())
-                    .setWindowingStrategyId(mainInputPCollection.getWindowingStrategyId())
-                    .setUniqueName(
-                        generateUniquePCollectonName(
-                            mainInputPCollection.getUniqueName() + "/PairWithRestriction",
-                            existingComponents))
-                    .build());
-
-        String splitAndSizeOutCoderId =
-            generateUniqueId(
-                mainInputPCollection.getCoderId() + "/SplitAndSize",
-                existingComponents::containsCoders);
-        rval.getComponentsBuilder()
-            .putCoders(
-                splitAndSizeOutCoderId,
-                ModelCoders.kvCoder(
-                    pairWithRestrictionOutCoderId, getOrAddDoubleCoder(existingComponents, rval)));
-
-        String splitAndSizeOutId =
-            generateUniqueId(
-                mainInputPCollectionId + "/SplitAndSize", existingComponents::containsPcollections);
-        rval.getComponentsBuilder()
-            .putPcollections(
-                splitAndSizeOutId,
-                PCollection.newBuilder()
-                    .setCoderId(splitAndSizeOutCoderId)
-                    .setIsBounded(mainInputPCollection.getIsBounded())
-                    .setWindowingStrategyId(mainInputPCollection.getWindowingStrategyId())
-                    .setUniqueName(
-                        generateUniquePCollectonName(
-                            mainInputPCollection.getUniqueName() + "/SplitAndSize",
-                            existingComponents))
-                    .build());
-
-        String truncateAndSizeCoderId =
-            generateUniqueId(
-                mainInputPCollection.getCoderId() + "/TruncateAndSize",
-                existingComponents::containsCoders);
-        rval.getComponentsBuilder()
-            .putCoders(
-                truncateAndSizeCoderId,
-                ModelCoders.kvCoder(
-                    splitAndSizeOutCoderId, getOrAddDoubleCoder(existingComponents, rval)));
-        String truncateAndSizeOutId =
-            generateUniqueId(
-                mainInputPCollectionId + "/TruncateAndSize",
-                existingComponents::containsPcollections);
-
-        rval.getComponentsBuilder()
-            .putPcollections(
-                truncateAndSizeOutId,
-                PCollection.newBuilder()
-                    .setCoderId(truncateAndSizeCoderId)
-                    .setIsBounded(mainInputPCollection.getIsBounded())
-                    .setWindowingStrategyId(mainInputPCollection.getWindowingStrategyId())
-                    .setUniqueName(
-                        generateUniquePCollectonName(
-                            mainInputPCollection.getUniqueName() + "/TruncateAndSize",
-                            existingComponents))
-                    .build());
-
-        String pairWithRestrictionId =
-            generateUniqueId(
-                transformId + "/PairWithRestriction", existingComponents::containsTransforms);
-        {
-          PTransform.Builder pairWithRestriction = PTransform.newBuilder();
-          pairWithRestriction.putAllInputs(splittableParDo.getInputsMap());
-          pairWithRestriction.putOutputs("out", pairWithRestrictionOutId);
-          pairWithRestriction.setUniqueName(
-              generateUniquePCollectonName(
-                  splittableParDo.getUniqueName() + "/PairWithRestriction", existingComponents));
-          pairWithRestriction.setSpec(
-              FunctionSpec.newBuilder()
-                  .setUrn(PTransformTranslation.SPLITTABLE_PAIR_WITH_RESTRICTION_URN)
-                  .setPayload(splittableParDo.getSpec().getPayload()));
-          pairWithRestriction.setEnvironmentId(splittableParDo.getEnvironmentId());
-          rval.getComponentsBuilder()
-              .putTransforms(pairWithRestrictionId, pairWithRestriction.build());
-        }
-
-        String splitAndSizeId =
-            generateUniqueId(transformId + "/SplitAndSize", existingComponents::containsTransforms);
-        {
-          PTransform.Builder splitAndSize = PTransform.newBuilder();
-          splitAndSize.putInputs(mainInputName, pairWithRestrictionOutId);
-          splitAndSize.putAllInputs(sideInputs);
-          splitAndSize.putOutputs("out", splitAndSizeOutId);
-          splitAndSize.setUniqueName(
-              generateUniquePCollectonName(
-                  splittableParDo.getUniqueName() + "/SplitAndSize", existingComponents));
-          splitAndSize.setSpec(
-              FunctionSpec.newBuilder()
-                  .setUrn(PTransformTranslation.SPLITTABLE_SPLIT_AND_SIZE_RESTRICTIONS_URN)
-                  .setPayload(splittableParDo.getSpec().getPayload()));
-          splitAndSize.setEnvironmentId(splittableParDo.getEnvironmentId());
-          rval.getComponentsBuilder().putTransforms(splitAndSizeId, splitAndSize.build());
-        }
-
-        String truncateAndSizeId =
-            generateUniqueId(
-                transformId + "/TruncateAndSize", existingComponents::containsTransforms);
-        {
-          PTransform.Builder truncateAndSize = PTransform.newBuilder();
-          truncateAndSize.putInputs(mainInputName, splitAndSizeOutId);
-          truncateAndSize.putAllInputs(sideInputs);
-          truncateAndSize.putOutputs("out", truncateAndSizeOutId);
-          truncateAndSize.setUniqueName(
-              generateUniquePCollectonName(
-                  splittableParDo.getUniqueName() + "/TruncateAndSize", existingComponents));
-          truncateAndSize.setSpec(
-              FunctionSpec.newBuilder()
-                  .setUrn(PTransformTranslation.SPLITTABLE_TRUNCATE_SIZED_RESTRICTION_URN)
-                  .setPayload(splittableParDo.getSpec().getPayload()));
-          truncateAndSize.setEnvironmentId(splittableParDo.getEnvironmentId());
-          rval.getComponentsBuilder().putTransforms(truncateAndSizeId, truncateAndSize.build());
-        }
-
-        String processSizedElementsAndRestrictionsId =
-            generateUniqueId(
-                transformId + "/ProcessSizedElementsAndRestrictions",
-                existingComponents::containsTransforms);
-        {
-          PTransform.Builder processSizedElementsAndRestrictions = PTransform.newBuilder();
-          processSizedElementsAndRestrictions.putInputs(mainInputName, truncateAndSizeOutId);
-          processSizedElementsAndRestrictions.putAllInputs(sideInputs);
-          processSizedElementsAndRestrictions.putAllOutputs(splittableParDo.getOutputsMap());
-          processSizedElementsAndRestrictions.setUniqueName(
-              generateUniquePCollectonName(
-                  splittableParDo.getUniqueName() + "/ProcessSizedElementsAndRestrictions",
-                  existingComponents));
-          processSizedElementsAndRestrictions.setSpec(
-              FunctionSpec.newBuilder()
-                  .setUrn(
-                      PTransformTranslation.SPLITTABLE_PROCESS_SIZED_ELEMENTS_AND_RESTRICTIONS_URN)
-                  .setPayload(splittableParDo.getSpec().getPayload()));
-          processSizedElementsAndRestrictions.setEnvironmentId(splittableParDo.getEnvironmentId());
-          rval.getComponentsBuilder()
-              .putTransforms(
-                  processSizedElementsAndRestrictionsId,
-                  processSizedElementsAndRestrictions.build());
-        }
-
-        PTransform.Builder newCompositeRoot =
-            splittableParDo
-                .toBuilder()
-                // Clear the original splittable ParDo spec and add all the new transforms as
-                // children.
-                .clearSpec()
-                .addAllSubtransforms(
-                    Arrays.asList(
-                        pairWithRestrictionId,
-                        splitAndSizeId,
-                        truncateAndSizeId,
-                        processSizedElementsAndRestrictionsId));
-        rval.setPtransform(newCompositeRoot);
-
-        return rval.build();
-      } catch (IOException e) {
-        throw new RuntimeException("Unable to perform expansion for transform " + transformId, e);
-      }
-    }
   }
 }