+++ sdks/python/apache_beam/pvalue.py
@@ -659,3 +659,27 @@ class Row(object):
   """
   def __init__(self, **kwargs):
     self.__dict__.update(kwargs)
+
+  def __iter__(self):
+    for _, value in sorted(self.__dict__.items()):
+      yield value
+
+  def __repr__(self):
+    return 'Row(%s)' % ', '.join(
+        '%s=%r' % kv for kv in sorted(self.__dict__.items()))
+
+  def __hash__(self):
+    return hash(type(sorted(self.__dict__.items())))
+
+  def __eq__(self, other):
+    return type(self) == type(other) and self.__dict__ == other.__dict__
+
+  def __ne__(self, other):
+    return not self == other
+
+  def __reduce__(self):
+    return _make_Row, tuple(sorted(self.__dict__.items()))
+
+
+def _make_Row(*items):
+  return Row(**dict(items))